+++ js/main.js
@@ -1,7 +1,7 @@
 import './util.js';
 import './data.js';
 import './miniatures.js';
-import './size-full-images.js';
+import './big-picture.js';
 //import './user-form.js';
